refactor: change structure and refactor api logic
