+++ frontend/src/app/app.module.ts
@@ -13,13 +13,13 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientModule } from '@angular/common/http';
-import { LoginComponent } from './login/login.component';
-import { VideosListComponent } from './videos-list/videos-list.component';
-import { VideoComponent } from './video/video.component';
-import { LayoutComponent } from './layout/layout.component';
-import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
+import { LoginComponent } from './pages/login/login.component';
+import { LayoutComponent } from './pages/layout/layout.component';
+import { PageNotFoundComponent } from './pages/page-not-found/page-not-found.component';
 import { FormsModule } from '@angular/forms';
-import { RegistrationComponent } from './registration/registration.component';
+import { VideosListComponent } from './pages/videos-list/videos-list.component';
+import { VideoComponent } from './components/video/video.component';
+import { RegistrationComponent } from './pages/registration/registration.component';
 
 @NgModule({
     declarations: [
