+++ frontend/src/app/components/video/video.component.html
@@ -1,4 +1,4 @@
 <span>{{ video._id }}: {{ video.title }}</span>
 <video autoPlay controls width="200">
-    <source [src]="apiUrl + video._id" type="video/mp4" />
+    <source [src]="videoUrl" type="video/mp4" />
 </video>
