+++ frontend/src/app/config.ts
@@ -1,0 +1,1 @@
+export const apiUrl = 'http://localhost:3002/api/v1';
