+++ frontend/src/app/http-interceptors/index.ts
@@ -1,6 +1,6 @@
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 
-import { AuthInterceptor } from './../auth.interceptor';
+import { AuthInterceptor } from './auth.interceptor';
 
 
 export const httpInterceptorProviders = [
