+++ frontend/src/app/pages/layout/layout.component.ts
