+++ frontend/src/app/pages/registration/registration.component.scss
