+++ frontend/src/app/pages/registration/registration.component.spec.ts
