+++ frontend/src/app/pages/videos-list/videos-list.component.scss
